fix(scripts): remove importer script and consolidate utility functions
