+++ .github/boring-cyborg.yml
@@ -19,7 +19,7 @@ firstPRWelcomeComment: >
   Thanks for opening this pull request! Please check out our contributing guidelines.
 
 verifyTitles:
-  titleRegexp: ^(feat|feat!|fix|fix!|docs|refactor|refactor!|test|chore)\((release|docs|deps|deps-dev|awesome:projects|ci)\):.*
+  titleRegexp: ^(feat|feat!|fix|fix!|docs|refactor|refactor!|test|chore)\((release|docs|deps|deps-dev|scripts|awesome:projects|ci)\):.*
   alwaysUsePrTitle: true
   validateEitherPrOrSingleCommitTitle: true
   statusTitle: 'Title Validator'
